date: clarify timezone handling of tm_to_time_t()
The point of this function is to convert a broken-down "struct tm" time
into a seconds-since-epoch time_t. We can only do this for UTC times,
since we don't know anything about zone handling (and if we did want to
respect the local zone, the system mktime() already does it for us;
there's no standardized alternative for UTC).

Signed-off-by: Jeff King <peff@peff.net>