date: allow @-timestamp dates to be negative
When parsing an input date from $GIT_AUTHOR_DATE, etc, we allow "@"
timestamps to specify a unix time in seconds since the epoch. Let's
teach that to allow negative dates, which makes it possible to create
objects with older dates in the usual tools like git-commit.

There should be little risk of ambiguity here, since the point of the
"@" format is to specify a unix time exactly. We can look at loosening
other more human-friendly formats in a future patch.

Note that we need to switch our error checking for parse_timestamp()
away from checking TIME_MAX, since we might get TIME_MIN. We can just
clear and check errno for ERANGE, which covers both cases.