rebase: exit when checkout fails
Our "checkout && rebase" chain avoids running "rebase" when checkout
fails. But it means we never hit the loop exits inside the rebase
conditional. And in most cases, a failed checkout will fail for _every_
branch (because it's caused by a dirty working tree or similar). Even if
it didn't, it makes sense to stop and manually assess the situation.

It's also not intentional. This is bug was accidentally introduced by
a3d9971 (rebase: auto-continue after complete rerere resolution,
2012-07-23), when we switched away from an &&-chain with an exit at the
end.