rebase: tighten auto-continue check
We'll try to auto-continue a rebase if there are no unmerged entries
left. This works OK if the reason the rebase failed was due to conflicts
(and they were either auto-resolved or not).

But if it fails for other reasons (e.g., a dirty working tree), the
auto-continue doesn't help. It will generally fail itself, but it
creates extra noise while doing so. Let's confirm a few more things
about the expected state:

  - there should be actual changes staged in the index to commit

  - there shouldn't be anything left unstaged in the working tree

These aren't completely foolproof (we could have bailed due to a dirty
index), but would catch common cases. The worst case is that we try to
continue and again fail.