topics: limit subject alignment / truncation
When using "topics -v", it's nice for the subjects to line up, and to
not be too long. We do this unconditionally, even though the script has
grown to have other output formats. In rare cases this can lead to some
amazing confusion. E.g., Meta/extra asks for a list of topics and their
sha1s. If a topic name is over 40 bytes, then the sha1 will be
truncated!

Let's add a "-p" option for pretty output, and turn it on by default
when using "-v".