private: abort on unknown todo instruction
If we see an unknown line in the todo list, our "case" doesn't match
anything and we just silently ignore it. This doesn't happen under
normal circumstances (since we're generating the list in the first
place), but can happen when manually editing (e.g., I tried to add a
"break" instruction, but we don't support it).

Aborting is the defensive thing to do, and also makes my "break" case
just work.