config.mak: stop converting SANITIZE into compiler options
This has been in the upstream Makefile since 56b5db30d0 (Makefile: add
helper for compiling with -fsanitize, 2017-07-10), so we're just
duplicating the options (which fortunately does nothing).

This does mean I won't be able to compile very old commits with
SANITIZE, but that seems like an unlikely scenario (and of course
tweaking CFLAGS manually is always an option).