config.mak: drop MSan config
I added this when playing around with MSan in 1daf114 (config.mak: turn
on address tracking for MSAN, 2015-02-14). But it has never actually
been a useful tool, and I haven't tried running it in years.

The last attempt was:

  https://lore.kernel.org/git/20171004101932.pai6wzcv2eohsicr@sigill.intra.peff.net/

which did yield some useful results, but requires annotating the code
with "unpoison" calls for data coming from libraries.