rebase: don't use HEAD in find_base
Since a986896 (rebase: let git-rebase checkout the branch, 2020-08-14),
we will run our find_base function without actually being on the branch
of interest. It should be looking at the difference between $upstream
and the topic branch, not HEAD.

Interestingly, this mostly worked anyway, because the upstream commit
is the same for most branches anyway (the most recent tip of master).
But:

  - it didn't actually handle the rebased-upstream magic that was the
    whole point of find_base in the first place (I suspect I didn't
    notice because it really only matters for dependent topics, and I
    haven't had any in the past few months)

  - if Meta/rebase was run from an older topic, chaos could ensue
    (trying to rebase a bunch of commits from the upstream as part of
    the topic)