rebase: drop --rebase-on-upstream option
I haven't used this in years. And in fact I had to use blame to even
figure out what it's for. It was added in ea6abed (rebase: add
rebase-on-upstream flag, 2013-12-22) as a workaround for when the
fork-point algorithm in find-base gets confused. That doesn't seem to
really happen. And in practice if I did run into a problem, I'd probably
manually run git-rebase on the branch in question.

It was also likely broken since a986896 (rebase: let git-rebase checkout
the branch, 2020-08-14), since it just reports "@{u}" (and it really
ought to be "topic@{u}".