config.mak: bump SPATCH_BATCH_SIZE
I noticed that there's a new version of coccinelle packaged in Debian,
so I timed it with various batch sizes. Looks like 8 is the winner, both
in runtime and CPU time.

Note that this is using "make -j16 coccicheck", so we have multiple
processes running at once. Supposedly spatch can work in parallel
itself, but that doesn't seem to be the case (even with "--jobs", it
always pegs exactly one core).