gitconfig: use https:// to access Junio's remote
This is all public and anonymous, so git:// works, but https is much
more secure (because I know I'm actually talking to GitHub, and not a
MiTM TCP hijacker).

I suspect I started this at git:// many years ago because http support
back then was lousy, plus it could actually be slower (more
round-trips). These days there are the same number of round trips at the
Git level, though probably a bit more for the TLS setup. However, since
I'm quite close to the nearest termination site (~10ms), it doesn't seem
to make a big difference (hyperfine says 1.2x to run ls-remote, but this
is an absolute difference of only 50ms or so).