test: let "-r" set GIT_META_TEST
This is much less confusing than having to set the environment variable
manually. It might be nice if we simply let any non-option arguments
turn into the test, but at least for now the upstream branch occupies
that spot. I feel like I rarely use that, though, so maybe it would be
better as "-u" or something.