test: add "-u" option for upstream
This replaces the use of any unknown arguments to specify the upstream
to rebase. It's rare for me to use this feature, and I'd like to make
way for something more useful for default arguments.

Note that technically we were passing all unknown options to rebase. But
in practice I don't think I ever passed anything but an upstream (maybe
"--root", but you can still do "-u --root" here).