test: turn cmdline arguments into test script
This allows the more natural:

  Meta/test make foo.o

instead of:

  GIT_META_TEST='make foo.o' Meta/test

Note that even though we could get the command arguments correctly
separated from "$@", we have to splat it all into a single string to be
passed to "rebase -x" anyway.

If you specify something nonsensical like:

  Meta/test -c make foo

we'll quietly ignore the extra arguments. Possibly we should warn, but
I'm not sure I care enough.