test: rename $GIT_META_TEST to just $cmd
The current name reflects its roots as an environment variable which we
took as input, and passed to a sub-process. But now it's just an
internal variable, so let's make that clear.

Technically this does disallow:

  GIT_META_TEST='make foo' Meta/test

but there's no need for that now that we accept the script on the
command line.