topics: use "grep -E" instead of egrep
As of gnu grep 3.8, using "egrep" generates an annoying warning. This
warning has been disabled by Debian (and I wrote this patch in the week
before they did). But it's probably worth switching over, as I'm not
sure of the ultimate fate of the warning (and it's possible I might even
have a non-Debian system some day).