merged: stop using upstream for merge checks
The current code uses the upstream as a "limit" for git-cherry. But this
produces confusing results when I actually do have an upstream that
isn't just "origin/master" (e.g., when I'm building on somebody else's
topic branch).

In that case as soon as the maintainer picks up the patches, then I
rebase mine on top of the upstream branch, and my branch becomes "empty"
when compared to the upstream. And so we say "oh, it's merged to master"
even though it clearly is not.

Perhaps it would make more sense to do:

  check_merge $i $upstream

in that case, and show when the patches were merged to the topic branch.
In theory if they're on somebody else's branch now, the maintainer will
deal with it. But I like to keep an eye on them flowing through.

The other usual upstream is "maint", but since it's a subset of "master"
in most cases, the limit doesn't do much anyway.

So let's just drop all of this upstream handling. It's not helping
anything and it makes the script slower.