merged: s/pu/seen/
The "--pu" option has been broken for a while, since there is no
"origin/pu" branch. Arguably this could just go away, as I don't seem to
actually use it (I didn't even notice the breakage from running it, but
just because I happened to be touching nearby lines). But it's easy
enough to port it forward. I'll give it another chance and see if I
remember to make use of it.