index_mem(): check return value of fsck_finish()
This would catch problems in blobs that were newly linked to git
metafiles (like .gitmodules) by a tree we are checking. We should add a
test.

Note that fsck_finish() both checks the blobs and frees the memory. So
in a short-circuit "||", we'd fail to do that cleanup if fsck_buffer()
reports a problem. That's OK in this case because we die(). Maybe this
should stay as two separate conditionals.