test: use getopt to canonicalize options
I don't love this, because it means repeating all of the options twice.
But I typed "-t1234" for the millionth time today and got annoyed (not
only does it not do what you want, but then you're in a broken rebase
that has to be aborted).

Obviously parsing "-t*" would work, too, but this feels less hacky (and
works for other options, too).