t0600: fix misplaced top-level ref
We have a test that tries to manipulate the branch refs/heads/referrent,
mostly using either the fully qualified refname or via git-branch (which
implies refs/heads/). However, one call to update-ref uses the
unqualified name, meaning we were quietly creating ".git/referrent",
which was otherwise unused by the test.

Fix this to specify refs/heads/referrent.

I _think_ it actually doesn't affect the test outcome either way. The
point of the test is that expiring reflogs for "the_symref" should not
fail when we cannot lock "refs/heads/referrent" that it points to
(because we have created a fake .lock file for it). And that is true
even if the "referrent" file does not even exist. After all, the process
holding the lock could be in the act of creating it.

So I suspect this "update-ref" line could just be dropped entirely.
Which you can verify by going back to its origin in 41d796ed5c (refs:
on symref reflog expire, lock symref not referrent, 2016-04-07) and
removing it. The test fails without the matching code change and passes
with it.

But I think it's worth keeping the update-ref call, as it creates a
situation which is more likely to match what we'd see in the real world.

Signed-off-by: Jeff King <peff@peff.net>