t5619: use fully qualified refname for branch
We use test_commit_bulk to update the ref "main". But without
qualification, that is really creating ".git/main", not an actual branch
in "refs/heads/".

In the end it does not really matter, because after creating "main" its
only purpose is for us to point HEAD to its same commit id. And that
works regardless of how we named it (the later call to update-ref just
calls it "main", but that is OK; we are resolving it to an oid there, so
our DWIM logic will find it in "refs/heads/").

Since it seems like going outside of "refs/" was accidental here, let's
do the more normal thing and just use a branch. That makes the test less
confusing and will future proof us against ref updates becoming more
picky.

Signed-off-by: Jeff King <peff@peff.net>