refs: check refnames as fully qualified when resolving
Most code paths for resolving refs end up in refs_resolve_ref_unsafe(),
which checks the names using check_refname_format(). The names we have
at this stage are always full refnames, but we pass the ALLOW_ONELEVEL
flag so that the function allows pseudorefs like MERGE_HEAD. We should
instead pass the FULLY_QUALIFIED flag, which lets check_refname_format()
do some extra syntactic checks on those pseudorefs.

With this patch we'll refuse to read anything outside of refs/ that does
not match the usual pseudoref syntax (all caps plus underscore). This
should not be a loss of functionality (since such refs cannot be written
as of the previous commit), but may protect us from mischief. For
example, you can ask for silly things like "rr-cache/<sha1>/postimage",
"info/refs", or "objects/info/commit-graphs/commit-graph-chain". It's
doubtful you can really do anything _too_ terrible there, but it seems
like peeking at random files in .git in response to possibly untrusted
input is something we should avoid.

Note that because we can't actually write such refs using Git, our tests
have to munge the filesystem manually to create a situation where
reading _could_ succeed. This only works with the files backend, of
course. When using other backends, we should still hit the same
restrictions, but they'd have failed anyway. Ideally we could realize
that the failure is because of a bad name, and not a missing entry, but
that's not currently possible. We don't return the "bad name" detail all
the way to the caller.

The second test here, for "main-worktree/bad", actually fails even
without this patch. The worktree-ref code enforces the pseudoref syntax
itself via is_current_worktree_ref(), so the extra checks for this in
check_refname_format() are redundant (though the parsing there is still
necessary so we know _not_ to reject "main-worktree/HEAD"). But either
way it's good to have a test which makes sure this remains the case.

Signed-off-by: Jeff King <peff@peff.net>