refs: check syntax of symref target
The symbolic-ref command checks that the target a symref points to is
syntactically well-formed. But it does so itself, and direct calls to
refs_update_symref() from other commands get no protection.

In theory this could allow you to do bad things, like pointing a symref
at refs/../../../some-file and reading the result. But in practice I
don't think this is too bad:

  - On the resolving side, we make sure that any targets we read are
    syntactically valid. So even if you can make such a bogus symref, it
    can't be used to read or write outside of refs/.

  - Looking over the other callers, I don't see any that can obviously
    be fed bogus input. They usually take the names of existing branches
    (e.g., to stick into HEAD). The biggest exception I found is "init
    --initial-branch", but it manually checks the validity of the branch
    name.

So I don't think there are any exploitable bugs here. But I think it's
better to enforce these rules within the ref code. That makes sure that
we haven't missed any cases. And even though we'd refuse to read the
refs, it's more helpful to detect the problem before even writing them.

Let's move the check from the symbolic-ref command into the refs code.
We can confirm via the existing tests that the bogus refnames are still
rejected.

However, we do need a few test updates:

  - I kept the same error text as the symbolic-ref message, but I
    lower-cased the first word to match our usual style. As a result,
    the test in t1401 needs to be updated to match.

  - A few tests (especially in t1430) use "test-tool ref-store" to
    intentionally create symrefs pointing to bogus targets. We can pass
    a flag there to disable verification. There's an interesting
    question here: which flag do we use? There is a flag for
    SKIP_REFNAME_VERIFICATION, but it is about the refname we are
    writing, not the destination. There is also SKIP_OID_VERIFICATION,
    which loosens checks on the destination. But of course there is no
    oid here at all, just a target refname.

    I went with SKIP_OID_VERIFICATION, as that seemed to be the moral
    equivalent for symrefs. We could add a new flag, but I don't think
    it's worth doing so. None of this is user-facing (in fact, most of
    these flags aren't even triggerable by real Git commands, and are
    only for the ref-store test-tool).